Add logic to instantiate classes.
And move class Gravity_Settings from aga-gravity-settings.php.
@todo: Instantiate more plugins in the function.
